+++ CHANGES.md
@@ -57,6 +57,7 @@ Release 1.7.3
 =============
 ### New Features ###
 * [PR#127](https://github.com/kaitoy/pcap4j/pull/127): Add support for setting immediate mode.
+* Add `isUp()` and `isRunning()` to PcapNetworkInterface.
 
 ### Bug Fixes ###